feat(Badge): 배지 찾는 메소드 작성
